pyproject.toml
[project]
name = "exammaker"
version = "0.1.0"
description = "EM – ExamMaker is an application designed for educators to simplify the process of creating and managing academic assessments."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml",
]

[project.scripts]
em = "exammaker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/exammaker"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
